requirements.txt
langchain==0.1.9
langchain-community==0.0.21
langchain-core==0.1.26
langchain-google-genai==0.0.9
langchain-pinecone==0.0.3
langsmith==0.1.0
pinecone-client==3.0.2
streamlit==1.31.1